Export .asperite file as .json
Export .aseprite file as .json and then in build step parse that .json and generate ANIMATION_FRAMES for every animation in array of .json files.
In build.rs we provide array of paths to .json files and then we get output for them in form of name, frame_count.

Simple and will be extended later.